+++ include/ess_format.h
@@ -35,45 +35,55 @@
 typedef enum ess_format {
    ESS_FORMAT_MONO_44100_8 =                             0 , /**< samplerate 44100   8 Bit Mono */
    ESS_FORMAT_MONO_44100_16 =                            1 , /**< samplerate 44100 16 Bit Mono */
-   ESS_FORMAT_MONO_44100_24 =                          2 , /**< samplerate 44100 24 Bit Mono */
+   ESS_FORMAT_MONO_44100_FLOAT_32 =                          2 , /**< samplerate 44100 32 Bit Mono */
    ESS_FORMAT_MONO_48000_8   =                          3 , /**< samplerate 48000   8 Bit Mono */
    ESS_FORMAT_MONO_48000_16 =                          4 , /**< samplerate 48000 16 Bit Mono */
-   ESS_FORMAT_MONO_48000_24 =                          5 , /**< samplerate 48000 24 Bit Mono */
+   ESS_FORMAT_MONO_48000_FLOAT_32 =                          5 , /**< samplerate 48000 32 Bit Mono */
    ESS_FORMAT_MONO_96000_8  =                           6 , /**< samplerate 96000   8 Bit Mono */
    ESS_FORMAT_MONO_96000_16 =                          7, /**< samplerate 96000 16 Bit Mono */
-   ESS_FORMAT_MONO_96000_24 =                          8, /**< samplerate 96000 24 Bit Mono */
+   ESS_FORMAT_MONO_96000_FLOAT_32 =                          8, /**< samplerate 96000 32 Bit Mono */
    ESS_FORMAT_STEREO_44100_8 =                          9 , /**< samplerate 44100   8 Bit Stereo */
    ESS_FORMAT_STEREO_44100_16 =                      10 , /**< samplerate 44100 16 Bit Stereo */
-   ESS_FORMAT_STEREO_44100_24 =                      11 , /**< samplerate 44100 24 Bit Stereo */
+   ESS_FORMAT_STEREO_44100_FLOAT_32 =                      11 , /**< samplerate 44100 32 Bit Stereo */
    ESS_FORMAT_STEREO_48000_8 =                        12 , /**< samplerate 48000   8 Bit Stereo */
    ESS_FORMAT_STEREO_48000_16 =                      13 , /**< samplerate 48000 16 Bit Stereo */
-   ESS_FORMAT_STEREO_48000_24 =              14, /**< samplerate 96000 24 Bit Stereo */
+   ESS_FORMAT_STEREO_48000_FLOAT_32 =              14, /**< samplerate 96000 32 Bit Stereo */
    ESS_FORMAT_STEREO_96000_8  =               15, /**< samplerate 96000   8 Bit Stereo */
    ESS_FORMAT_STEREO_96000_16 =              16, /**< samplerate 96000 16 Bit Stereo */
-   ESS_FORMAT_STEREO_96000_24 =              17, /**< samplerate 96000 24 Bit Stereo */
+   ESS_FORMAT_STEREO_96000_FLOAT_32 =              17, /**< samplerate 96000 32 Bit Stereo */
 
    ESS_FORMAT_MAX  = 18,
 } ess_format_t;
 
 /**
  * @brief Help to get the number of channels of the format.
- * @param format The format to parse
+ * @param [in] format The format to parse
+ * @return channels of the format
  */
 int ess_format_get_channels(const ess_format_t format);
 /**
  * @brief Help to get the samplerate of the format.
- * @param format The format to parse
+ * @param [in] format The format to parse
+ * @return samplerate of the format
  */
 int ess_format_get_samplerate(const ess_format_t format);
 /**
  * @brief Help to get bits of the format.
- * @param format The format to parse
+ * @param [in] format The format to parse
+ * @return bits of the format
  */
 int ess_format_get_bits(const ess_format_t format) ;
 /**
  * @brief Help to get the string of the format.
- * @param format The format to parse
+ * @param [in] format The format to parse
+ * @return string  of the format
  */
 const char* ess_format_to_string(const ess_format_t format);
 
+/**
+ * @brief Help to get the sample size of the format.
+ * @param format The format to parse
+ * @return samplesize of the format
+ */
+int ess_format_get_sample_size(const ess_format_t format);
 #endif